Refactor breakdown API and UI for enhanced exam result analysis.
- Improved the `GET /api/exams/{exam_id}/breakdown` endpoint to deliver a more detailed breakdown of exam results by student and concept.
- Updated the ExplainPage in the frontend for better visualization of breakdown data, aiding teachers in reviewing judgments and rubric adherence.
- Revised documentation in README.md to include new features and functionalities related to the breakdown API and UI enhancements.
- Optimized data structures in the frontend to ensure robust handling of breakdown information for student evaluations.

+++ scales_v3/scales/models/result.py
@@ -20,7 +20,7 @@ class ConceptFeedback(BaseModel):
     knowledge_point: str
     verdict: Verdict
     marks_awarded: float = Field(..., ge=0)
-    max_marks: int = Field(..., gt=0)
+    max_marks: float = Field(..., gt=0)
     evidence_span: str = Field(default="")
     reasoning: str = Field(default="")
     trust_score: float = Field(..., ge=0.0, le=1.0)
